fix(CT1) remove query from loop
This commit fixes a performance issue where the same query would run in
a loop slowing event insertion down quite a bit.

Furthermore, this commit add the `Builder::all()` method to fetch all
the results for a query in a batched `Generator`.
Also: allow controlling the output format of the results: `OBJECT` will
return an instance of the model, `ARRAY_A` and `ARRAY_N` will return the
results in array format the same way the `wpdb::get_results()`  method
works.