fix(CT1) do not reuse first Occurrence after loop 1
When updating Occurrences, if no match is found, then we try to reuse
the first occurrence to make sure Single Events will update "in-place"
and the Occurrence `occurrence_id`, and thus provisional ID, will not
change over updates.
This fix makes sure the first occurrence is only used for a possible
match in the first run of the loop generating, or updating, Occurrences.