fix(CT1/Models/Occurrence) better first occurrence match
TEC logic to upsert Occurrences will run in the context of ECP too.
When dealing with Single Events, then the event should be updated
in-place, to make sure that Single Events will have their Occurrence
`occurrenc_id`, thus provisional post ID, not change across updates.

In the context of ECP, when updating Recurring Events, instead,
Occurrences should not "move" (have their date and time changed) to
match the new required set, but should be removed and recreated if not
matching the new pattern.

This fix will provide the logic filtering the
`tec_custom_tables_v1_get_occurrence_match` filter the first occurrence
object **only on the first run of the loop** and will allow filtering
logic to null it.

ECP will null it if not matching any new required Occurrence.