fix(CT1/WP_Query) apply modified to AJAX requests
This fix ensures the `Events_Only_Modifier` will apply to AJAX requests,
as it should.
The previous version of the code was running an `is_admin()` check to
avoid filtering queries in the context of the admin UI, but `is_admin()`
would return `true` even in the context of requests to the
`admin-ajax.php` endpoint.