refactor(CT1/Models) faster update
This commit fixes the issue where updates of existing information would
be slow.
While insertions would be make in bulk, the updates would be made
Occurrence by Occurrence.
With never-ending daily Events, this would mean an All update to a
Recurring Event end time would entail some 3650 queries.
There are 2 slow parts in the process:
1. The queries will run 1 by 1, each paying the over-head of a TCP
   communication and operation.
2. Any update is, first, a SELECT on the whole table on the (indexed,
   still...) primary key.

This commit DELETEs the rows to update and re-INSERTs them in batches
bringing the update time in line with the INSERT time.
Why does this work? Since we use the primary key as part of the
re-inserted data, the re-insertion will not alter the data.