build(composer.json) add action-scheduler
Require version 3.4.0 or above.
Action Scheduler comes bundled with WooCommerce, but the possible
conflict of double-loading and version conflict is handled with a system
similar to the one we use in Common when bundled in TEC and ET: the most
recent version will load. If TEC is installed and activated on a site
using WooCommerce, then it will not be an issue.