fix(CT1/Migraiton) handle die/exit in Worker
Depending on the PHP version and runtime (php-fpm vs cli vs Apache
module) the functions hooked on shutdown might not be called when the
code calls teh `die` or `exit` function.
To cope with this this commit leverages the fact that PHP is, first and
foremost, a scripting language to emit HTML code and, as such, will
outptut whatever output the script produced before closing the VM.
This means that callbacks hooked to the `ob_start` function will always
be called when PHP flushes the buffer.
The worker leverages this by setting a flag property,
`Process_Worker::migration_completed` to know if a flush happens before
the worker itself calls it or not.
If called before, then the Worker will correctly flag the migration as
failed and will log whatever output the script produced before the `die`
or `exit` call.