fix(CT1/Migration) init State from bad option
This commit fixes the `Migration/State` code to make sure it would
correctly initialize from "bad" option values.
The content, and format, of the option should not be trusted and someone
might set the content of the Migration state sub-option to `null` and
expect that to reset the Migration state. That would, before this
commit, fail and result in a fatal.