fix(CT1/Migration) do not poll before start
This commit updates the Ajax handling class code to return, together
with the unchanged information, a `poll` flag key to indicate to the
frontend JS code whether it should keep, or start, polling the backed
for updates or not.
When the user is looking at the Start and Complete screens of the
Preview, Migration and Undo phases, then no AJAX requests should be
polling the backend for information that is, to change, waiting on the
user interaction.