fix(CT1/Migration) handle migration race conditions
Modify the phase check to read like "there are no Events that need
migration AND there are no Events that are being migrated".
In the worker, when there are still in progress migration, queue an
action to check on the migration state that will keep enqueueing a new
instance of itself until done.