refactor(CT1/Migration) remove cached values
The state will be queried at the start of the operations and at the end,
a cached value at the start might lead to a wrong when hitting that same
cached value at the end.
This commit changes the various `get_total...` methods to be "live"
methods that will query the database on each call and will not return
cached values.
I've, first, gone for a `$force = false` parameter to transmit down from
the callee to each called function, and found out we'd mostly call the
functions using `$force = true`. This would make the decision about what
to cache when `$force` parameter is `false` prone to error and would
introduce other consistency issues.
Since we were caching on a per-request basis using the `Tribe__Cache`
object by means of its `ArrayAccess` API, we would get little benefit
from caching, a couple queries in a query intensive context, with a
trade-off of correctness.