refactor(CT1/Migration) check_phase in Process_Worker
This commit adds more phase checking steps to the `Process_Worker` to
avoid queueing a phase checking action when not required.
This change follows the one made to make the `get_total` functions not
caching and "live", thus enabling pre and post checking of the new,
updated, database status in the context of a single worker.