fix(CT1/Migration) check the phase on exit
By means of the method hooked, in the `Process_Worker`, to the flushing
of the output, we get a last chance to do something. If the method logic
is running, then the Event migration was not completed and _something_
called `die` or `exit`. It's, then, unlikely the migration will be able
to queue any more actions. For the sake of safety, call the idem-potent
`check_phase` method and transit the migration phase, if required and
while we can.