fix(Events/API) use timezone-sensitive timestamps
When calculating the `_EventDuration` meta in the context of the
`API::saveEventMeta` method, the function used would be the `strtotime`
one.
The problem with that, is the function will use the **system** timezone,
not the site one.
In the example crystalized in the tests, given daylight saving kicks in
on Sunday 2022/03/27 at 2 am in France, an event from 1:30 am to 4:30 am
that day would have an incorrect duration of 3 hours. At 2 am the clock
is moved to 3am and the Event would actually last, for someone at the
event, 2 hours, not 3.
The fix calculates the timestamp using PHP built-in `DateTime` and
`DateTimeZone` classes, through the util wrappers, to make sure the
duration is correctly calculated.