refactor(CT1/Migration) show Upg. tab if reuqired
This commit updates the logic that would hook on the display of the
Upgrade tab, from TEC Main code, to make sure the Upgrade tab will show
only when required.
The commit adds a possible state to the migration phase: the one where
the migration is not required at all due, i.e. there are no events on
the site to begin with.
The second part of the commit updates the logic in the `Upgrade_Tab`
class to hook on the `tec_events_upgrade_tab_has_content` filter and
builds upon the Migration `State` to show only if the migration is
required, running or has completed.