refactor(CT1/Migration) add a group to the table schema
This commit updates the table schema code to add a `group_name` method
to each table definition. This group will be used in the context of the
`Schema_Builder::all_tables_exist` method to either check all the
possible registered tables (unchanged from previous implementation) or
to only check a sub-set of the tables using the group filter.