fix(CT1) on activation create tables
This fix builds on the fix put in place for Common to make sure Common
would completely load on the plugin activation.
This fix ensures the tables will be created when the plugin is activated
(if it's activated) either by means of the Admin UI, or by means of
wp-cli.