fix(CT1/Migration) updates following transaction move
Updates to the code to support the move of the transactional logic from
the strategies to the `Process_Worker`.
The commit adds a class-wide query execution toggle to the `Builder`
class and make sure the Single Event migration strategy will work
correctly in preview mode when transactions are not supported.