Fix(API.php) daylight saving time crossing duration
This commit fixes an issue where the `API.php` method calculating the
duration of an event, the seconds value stored in the `_EventDuration`
meta value, would incorrectly handle daylight-saving times.

Explained with an example: in France daylight-saving time starts on
2022/03/27 at 2 AM. An Event with a start date of 2022/03/27 1 AM
(before the daylight saving time change) that ends on 2022/03/27 at 4 AM
would have a real duration of 2 hours, not 3 as 4 - 1 would suggest:

* 1 hour from 1 AM to 2 AM
* at 2 AM the daylight saving time moves the time from 2 AM to 3 AM
* 1 hour from 3 AM to 4 AM.

The bug is a bit remote and was hidden, so far, by the fact that most
views will use the Event start and end time, rather than duration, to
render. Additionally, an Event would have to cross the daylight saving
time to be affected.

The issue was caused by the code using the `date`, et al., functions to
calculate the timestamps; these are functions unaware of the daylight
saving time. Using timezone-aware `DateTimes` object brought the issue
up.

Due to the Views not relying on the Duration to render, I cannot
providee a screencast and am providing, among the changed files, the
next best thing: automated tests.