fix(Views/V2) suppress filters only on main query
Found while working on RBE and ECP code.
Views V2 is jumping in on the main query, the one that runs on PHP
initial state, to prevent the normal query logic from running it and to
inject the posts according to the parameters required by a View.

While meant to run on the main query only, the check was a bit too lax
and would not check if the filtered query is **actually** the PHP
initial state one.

This commit adds such check to avoid the query filtering suppression
from happening everywhere.
Furthermore, I've conflated an abstract class with its only extending
class to remove a bit of cruft and updated some code to avoid it from
running when no more required.