fix(CT1/Migration) exclude Events w/ post_parent
When an Event has a `post_parent` value that is not `0`, then it should
not be migrated.
In the context of ECP the `post_parent` field is used to relate
Occurrences of a Recurring Event to the original Event; the migration
should only deal with the original Event and not its children Events.