This fixes the application of the TOD filter in FBAR.
The problem: Events created in the block editor are filtered out by any TOD (time of day) filter.
The cause: Events created in the blocks editor do not have the _EventDuration meta set.
The solution: hook into event save and the model `build_properties()` calculate the duration and save it.

As part of this fix, I also discovered that somewhere Event->build_properties() is being called on a venue.
So I put in some sanity checks to prevent that.

Original fix. At the end you'll see a debugged duration of 0 - which lead me to the sanity checks.
:movie-camera: https://d.pr/v/L2LmxS