test(ct1_migration) backup hooks after bootstrap
The suite bootstrap file will register the CT1 provider that will, in
turn, add hooks.
The test case `setUpBeforeClass` method will run **before** the
bootstrap file is loaded and will backup a filter state that does not
contain the CT1 hooks; this commit fixes it be backing up the hooks to
be restored after each test after the bootstrap file is loaded, in the
test case `setUp` methd.