fix(CT1) suppress errors on failure to lock
While trying to get and lock the IDs of the Events to migrate, the code
supports and handles the case where the locking will fail. When an
`INSERT ... SELECT` query fails the MySQL driver will report an error.
In this case, suppress the errors to avoid the failure to lock appear as
an error in the logs.