fix(CT1) ensure consistent migration state
When the tables already exist and the plugin is reactivated, and this
might happen in the context of a plugin update by means of fetching or
zip uploading, then the state of the migration would not be set
correctly on empty sites.

This commit fixes that and adds a test to make sure the fix works
correctly.