fix(CT1) fully activate during plugin activation
The `Provider` code will run on `tribe_plugins_loaded` and will find the
state `null` and will not register the `Full_Activation_Provider` on an
empty site.
The `Activation` code will run later, on `init`, and will set the state
correctly.

This fix adds a check in that second step to make sure we're handling
that case and fully-activating the plugin in the context of that
request.