refactor(CT1) pull in data cache from ECP repo
This commit updates the code of the repository filtering logic to make
sure the ECP version of the repository filtering logic will not require
a dedicated `update` method to work correctly.

This allows to create one less wrapper on the ECP side and simplifies
the ECP creation and update logic to avoid having the Occurrence and
Series Relationship update logic running twice.