fix(CT1) Update page assets
When Event Ticktes is active, then ET own `tribe_resource_url` function
will kick in and will use, where an `$origin` is not provided, ET
resources directory as a source for the assets.

Making the `$origin` explicit will force the function to look to
the TEC resources directory for those assets.