doc(CT1/Migration) update messaging about rec. evts
When trying to migrate a Recurring Event using the Single Event
migration strategy we're, correctly, blocking the migration and putting
out a message about it.

Update the message of that kind of error to provide an actionable
solution: usually the issue is ECP is not active on the site during the
migration process.