fix(CT1/Migration) if evnt has no date info, skip
This commit updates the logic that is countin the Events to migrate and
the one that is getting and locking the IDs of the Events to migrate to
not include Events that have missing date-related information.

While the migration code can do quite a lot to handle reconstruct an
Event date-related informatio from what is there, an Event missing the
basic `_EventStartDate` and `_EventStartDateUTC` or `_EventEndDate` and
`_EventEndDateUTC` information cannot be migrated.

Those events are, usually either trashed events or artefacts of failed
imports that would leave a user in a situation where the only option is
manually deleting each one; with this fix, those events will be
completely ignored by the migration process as they are, most likely,
ignored by the user.