fix(CT1/Migration) do not assume ECP active
Do not assume ECP is active in warning handler.

The migration code has an error handler that will deal with `E_WARNINGS`
coming from plugins that are not TEC or ECP differently. That code has a
bug where the existence of the `EVENTS_CALENDAR_PRO_FILE` constant is
assumed.

This commit fixes that.