feat(REST API) suport strict_dates on /events
By default, unless the site administrator has filtered the
`tribe_events_rest_use_inclusive_start_end_dates` filter to return a
falsy value, a REST API request for the `/events` endpoint will cast the
`start_date` parameter to the start of the day, and the `end_date`
parameter to the end of the day.

This adds a `strict_dates` supported REST API argument to allow REST API
clients to control, in the context of a request, the default value of
the value filtered in the
`tribe_events_rest_use_inclusive_start_end_dates` filter.

The filtered value will still override this value if explicitly set.

Furthermore, I've fixed the description of the `end_date` parameter: it
will limit the fetch by Events ending before a date and time; it's not
about the start date.

# Conflicts:
#	readme.txt