fix(CT1) correct set properties of redirected query
When taking over a query in the context of CT1, set the `found_posts`
and `max_num_pages` properties correctly on the original query.

In CT1 context, a `WP_Query` for Events request will not run; in its
place will run a `Custom_Tables_Query` built on top of it that would
not, before this fix, correctly set all the expected properties on the
original `WP_Query` object.