fix(Integrations/Divi) do not reset postdata on ID
When trying to determine the post ID of an post that is not an Event,
the current code would call the `wp_reset_postdata` function.
This would reset the postdata to the post ID of the queries post, page
or whatever, invalidating any custom query that might run in the context
of the same page.