feature(CT1/Migration) piggyback on JS to run
The Migration UI poll the backend to fetch the current migration status
HTML.
This update hooks on that report request to migrate some non migrated
Events.
This processing on the back of the XHR request to the backend will start
by migrating the not yet migrated event and will, then start migrating
Events that have been claimed by Action Scheduler actions are are
pending (not yet started).

Since this processing works on the premise of the Migration UI polling
the backend, the code does not exclude Action Scheduler and will,
instead, rely on it should the user close the tab/window/browser.