fix(Query) do not apply date order to non-Events
This fixes an issue where the event-date based order, based on the
`_EventStartDate` or `_EventEndDate` meta, would be erroneously applied
to all post types, not just Events.

Notably, this would result in tickets related to Events not being
correctly displayed on the front-end.

The cause of the issue was the Query code would check hook the filtering
and ordering function on any query for a TEC post type. The detection,
in the Context component, of the post type is based on the **request**
variables.
So, a request to render the Event single would be a main query for a TEC
post type, any following query for any TEC post type (e.g. Venues or
Tickets) would then be filtered.

The fix consists in checking, per-query, if the query is for only Events
or not.