Move the filtering and translation of view labels to the View class.
In the View Manager class, we have been passing the label and text domain as variables. `__( $text, $domain )`

This does not work.
This changeset moves that handling to the View class itself - and deprecates the offending methods and filters.
This only requires the (non-abstract) subclasses (specific views) have a construct method that defines the label in the standard `__( string, string )` manner.

Exceptions:
 - Latest Past  - as I believe that inherits it's label from the original view.

Tests: none yet
Checks: I have generated a test .pot file and confirmed the missing strings are available for translation. (not included in the PR)

[TEC-4485]