fix(CT1) HyperDB plugin compatibility issue
This fixes an issue where, in the context of PHP 8.0, a fatal would be
thrown on Event archives when using CT1 and the HyperDB plugin.

The cause of the isues: in the context of the `SELECT
FOUND_ROWS()` query that runs when the query is instructed to set the
found rows (i.e. `no_found_rows` not set or set to `false`), the plugin
will try to access the `mysqli_result` immediately.
When the Custom Query injects the post in the main `WP_Query`, then the
`mysqli_result` instance the `SELECT FOUND_ROWS()` query runs on
would not be able to provide and an error would be generated in the
`wpdb` class.

In previous PHP version, the default behaviour on error would be to
continue, hence the `mysqli_result` instance would survive the error.

From PHP 8.0 and up, the default behaviour on error is to null the
`mysqli_result` instance, from here the error reported.

The solution is to filter the query that runs to set the `found_rows`
providing the result already built on the Custom Query.