fix(Query) correctly handle inclusion of Events in loop
This updates the logic that would add the Events post type to the query
post types to make sure main queries for the `post` type are correctly
handled.

Since the `post` type is the default one, WordPress will not specify the
`post_type` argument in the query and this would be incorrectly handled
in the `Tribe__Events__Query` class.