Update manager.js
The  "accepts: 'html'" is harmful and results in buggy behaviour depending on the server.   Because it is invalid in this form, it actually causes the Accept header to be "undefined", which may result in server side errors.

From http://api.jquery.com/jquery.ajax/ : 

accepts (default: depends on dataType)
Type: PlainObject
A set of key/value pairs that map a given dataType to its MIME type, which gets sent in the Accept request header. This header tells the server what kind of response it will accept in return. For example, the following defines a custom type mycustomtype to be sent with the request:

So, if the accepts needs to be specified, it should be on the form:
accepts: {
 'html': 'text/html'
}

This does not seem to add anything, and the KISS solution is to simply delete this line.