fix(CT1) filter Query Monitor implementations >=1
This fixes an issue emerged while working on the ACF integration where
additional Query Modifiers would not be filtered by the Query Monitor
due to an order of operations issue.
The missed filtering of the Query Modifiers would lead the ACF query
modifier to never being applied and cause, as a consequence,
malfunctioning of the ACF integration.