fix(Main) do not save meta for revisions
In WordPress 6.4, by default, revisions are saved **after** the post is saved.
Previously, the revision would be saved **before** the Event post.
Saving the revision before would not cause issues, but saving it after would lead to some of the Event meta, notably the recurrence meta, being set to an empty value.
This fix removes the function to save meta for revisions completely since meta saved for revisions is, anyway, always saved on the original post (see the `update_post_meta` function).