build(webpack.config.js) add handler for PostCSS nesting.
By default `@wordpress/scripts` will expect and compile to CSS nesting
syntax.

Given this PostCSS code:
```postcss
.selector{
  color: red;

  &--green{
    color: green;
  }
}
```

`@wordpress/scripts` would compile to:
```
.selector{
  color: red;

  & --green{
    color: green;
  }
}
```
That unrolls to:
```
.selector {
  color: red;
}
.selector --green{
  color: green;
}
```
While PostCSS nesting, provided by the `postcss-nested` plugin, will
compile to:
```
.selector {
  color: red;
}
.selector--green{
  color: green;
}
```
Which is the original intention of the PostCSS code.

This updates the schema for TEC PostCSS files to pre-process with the
`postcss-nested` plugin to, then, apply the rest of the PostCSS loaders
correctly.