test(classy_jest) add store mock facilities
I could not find a satsfying API to unregister a store and replace it in
the WordPress default registry for the purpose of testing. While I could
have updated the Classy components to use an HOC (e.g. `withRegistry`)
to inject either the default registry or a test one, I do not like
complicating production code to accomodate for testing. Enter store
mocking facilities. I've rolled my own as I could not find somethign
similar in the Gutenberg project.