fix(CT1) prevent too long query being killed
This adds the `tec_query_batch_size` function that returns a filtered
value representing the size of a chunk in the context of all batched
queries done by the plugin.
The value is then used in code that operates on unbounded sets to fetch
them honoring that batch size and avoiding the query to become too
large.