tweak(CT1) add indexes to occurrences table
This updates the Occurrence table schema to add two mult-column indexes:
* post_id, end_date, start_date
* post_id, end_date_utc, start_date_utc

The order of the columns matters in the indexes.
The `post_id` will always be part of the queryes in a JOIN since it
connects the occurrence to the event post.
The dates will always be used either in UTC format or not, and will not
be used in both formats in one query (the performance of which we care
about); this is the reason for the two indexes.
Finally, the `end_date` field comes before the `start_date` one as there
are slow queries that only use `end_date`, but there are no slow queries
(I currently know about) that use `start_date`, but not `end_date`.