tweak(CT1) introduce query caching in the Model builder.
This introduces query caching in the model builder to cache the results
of duplicated queries at the builder level without running them twice.

Proper cache invalidation was already directly
(`ct1_integration/BuilderTest.php`) and indirectly covered in existing
tests.

I had to modify a single test that was insisting on the batched fetching
functionality since the whole point of the test is to assert on the
number of queries made.