test(classy_jest) fix and update
Update some tests and the mock required for them to run.

Tests targeting the `EventDetails` and `TinyMceEditor` components are missing and I've skipped the test of the `Classy` component that would not use the `core/editor` store.

Those will be addressed in a follow-up PR to refactor the Classy approach to use a dedicated store registry and controls to communicate with the Core stores when available.