refactor(Classy) scaffold dedicated registry
This work starts the update of the Classy application to use a
dedicated, and separate from the WordPress Block Editor one, registry.

Why?

Since the application must work in the Block Editor context, and outside
of it, it cannot rely on the WordPress Core stores being present and
initialized on the page.

The app is using components (e.g. the `PostFeaturedImage` one) that have
hard-codeded dependencies on the Core stores (e.g. the `core` and the
`core/editor` stores); to intercept those dependencies, the only option
I could find is to create a new registry and, in that, re-route the
request to the correct core store in the core registry, if available.

At this stage the components render without crashing, but the featured
image one will not update correctly.

Tests will need update too to use a dedicated registry in place of store
mocks.