refactor(Classy) start move of TEC code back
There and back again.
This moves _some_ code from Common to TEC to start the move back of the
TEC-specific components previously moved to Common.
Included is the work to scaffold a first TEC Classy application
registered and enqueued following the Classy one in the PHP Controller
for the feature.