fix(Classy) React imports
For the IDEs (both Intellij and VSCode-like) to correctly pick up
components defined in the WordPress core components, React must be
imported using the syntax:
```
import * as React from 'react';
import {OtherStuff} from 'react'; // Other imports.
```

This updates the components that were missing this type of import to
reduce the number of TypeScript warnings.