test(classy_jest) connection to Common
This adds a first test that is currently doing nothing but works out the
connection from TEC to Common in the context of tests.

Specifically it allows the functions under test, `renderFields` to work
correctly when the source code of the function is importing functions
from Common using the `@tec/common` namespace.