refactor(Classy) remove todo
The code referenced in the todo is old code, and it would create a
dependency on an implementation that does not know about the use that is
done of it in the context of this feature. For this reason, I think it's
better to keep a curated list of meta keys in the context of the
feature.