fix(Classy) new event save
This fixes an issue where creating a new event would not save all the
information to the database required for it to correctly display on the
frontend.
The missing information was UTC information.
The fix consists in:
1. always attaching the start/end date/time to the saved information on
   first event save.
2. patching that information once it reaches the backend (in the Classy
   controller) to add the required UTC information if missing
3. letting the standard code follow

I've not updated the code to provide the UTC information in the JS code
to avoid the discrepancies that, in _some_ cases, exist between
timezones in JS and PHP.