fix(Classy) restre linked post save
This post restores the Venue create/edit functionality I had either not
pushed or that got lost in a merge.
The functionality is still buggy, though, and while venues and
organizers are correctly loaded, they seem to be not correctly saved in
the database resulting in array entries where there should be integer
entries.