fix(Classy) sanitize multi-meta correctly
While the meta might be multiple (i.e. not single or not unique), the
sanitization function will still be called for each single value of the
meta. E.g. for the multiple meta `_EventOrganizerID` field, when having
to save 3 values, the sanitization function will be called 3 times for
each one of the 3 integer values; not once on an array of 3 integers.