fix(Classy) build date objs correctly
The `getDate` function from the `@wordpress/date` package will use the
current site timezone to build the date, resulting in dates that are not
built on the `Z` (UTC) timezone and that will throw off the UI. The
`getEventDateTimeDetails` function should not apply the timezone to the
built objects since the dates are only used to move information around,
not manage timezone offsets. Managing the timezone offset is the backend
job.