fix(Classy) date handling in EventDateTime
The field will manage start date, end date, start time and end time.
It was doing so using the `getDate` function from the `@wordpress/date`
package that returns a `Date` object built using the current WP
timezone. This would throw a number of calculations about date/time off
when the timezone is not one aligned with UTC. Since, at the Event
level, the time zone concern is managed in the backend, the field should
not move around dates that contain any timezone information.