refactor(ORM) avoid TIMESTAMPDIFF if possible
This updates the repository method fetching overlapping events, to use
logic that will not require casting of the start / end dates and the use
of the `TIMESTAMPDIFF` function.

The method changed is the one serving this repository API:

```php
tribe_events()->where('date_overlaps', $start, $end);
```

By default the overlap interval is 1 second, i.e. a non-inclusive check,
but a different overlap can be specified:

```php
// Overlap of 23 seconds.
tribe_events()->where('date_overlaps', $start, $end, $timezone, 23);
```

If the specified overlap is more than `1`, then the function must use
the previous logic based on `TIMESTAMPDIFF`, else it will use a
string-based logic leveraging the fact that the start/end date/times are
saved in the format `Y-m-d H:i:s` that allows for string comparison
while making sense in numbers as well.