refactor(By_Day_View) date_overlap with 1
This change will make the repository code run avoiding the use of the
`TIMESTAMPDIFF` function.
I could not gather the original intention behind the commit that, in
2020, updated and moved this code, but I can guess the idea was to apply
a non-inclusive match by using an overlap of at least 2 seconds.
The same effect, but avoiding the `TIMESTAMPDIFF` function, can be
obtained using an overlap of 1 second.