test(classy_jest) resolve Common modules to the TEC modules
During tests the modules mocked in the tests will come from TEC own
`node_modules` directory, but components or functions defined in the
Common classy package that import modules will import those from the
`common/node_modules` directory. This is a problem as mocking a module
(e.g. the `@wordpress/data` one) in TEC test for a component that uses a
component/function defined in Common would not have any effect: the
module would be mocked for the TEC component, but it would not be mocked
in the Common component.