feat(Classy) localized end-of-day cutoff end h/m
To avoid having to repeat the calculation over and over in the Classy JS
code, calculate the end-of-day cutoff end hours and minutes once, in the
backend.

The values are ints like the end-of-day cutoff `hours` and `minutes`
values.
E.g.
```json
{
  "hours": 0,
  "minutes": 0,
  "endHours": 23,
  "endMinutes": 59,
}
```

The same applies to custom end-of-day cutoff settings like `2:30`:
```json
{
  "hours": 2,
  "minutes": 30,
  "endHours": 2,
  "endMinutes": 29,
}
```