refactor(Classy) createSelector to avoid useSelect warnings
This fixes an issue where using the selector in the context of a
`useSelect` hook would trigger the `useSelect` warning due to useless
repaints. While the content of the object returned by the selector would
be the same, the **reference** would not since the selector returns a
new object on each call. The memoization of the selector saves
computations and repaints when the actual data has not changed.